code/v1/requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
pandas
